refactor: move constants to a separate folder
