+++ src/main/kotlin/lotto/constants/LottoConstants.kt
@@ -1,3 +1,3 @@
-package lotto
+package lotto.constants
 
 const val TICKET_PRICE = 1000
